+++ src/base/addfld.f90
@@ -530,20 +530,21 @@ if (imeteo.ge.2) then
 
   ! Pure convection equation (no convection, no time term)
   call field_get_key_struct_var_cal_opt(iflid, vcopt)
+  vcopt%iwarni= 2
   vcopt%iconv = 0
   vcopt%blencv= 0.d0 ! Pure upwind
   vcopt%istat = 0
   vcopt%idircl = 1
-  vcopt%nswrsm = 100
+  vcopt%nswrsm = 5
   vcopt%nswrgr = 100
-  vcopt%imrgra = 0
-  vcopt%imligr = -1
-  vcopt%epsilo = 0.000001
-  vcopt%epsrsm = 1.d-3
+  vcopt%imrgra = 1
+  vcopt%imligr = 1
+  vcopt%epsilo = 0.00001
+  vcopt%epsrsm = 1.d-4
   vcopt%epsrgr = 0.0001
   vcopt%climgr = 1.5
   vcopt%idiff  = 1
-  vcopt%idifft = 0
+  vcopt%idifft = 1
   vcopt%idften = 1
   vcopt%relaxv = 1.d0 ! No relaxation, even for steady algorithm.
   vcopt%thetav = 1
@@ -559,6 +560,7 @@ if (imeteo.ge.2) then
   f_label = 'Meteo Temperature'
   ! Now create matching property
   call add_property_field(f_name, f_label, 1, .false., iflid)
+  call field_set_key_int(iflid, keylog, 1)
 
   f_name  = 'meteo_pot_temperature'
   f_label = 'Meteo pot Temperature'